pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "bmauth"
version = "0.3.0"
description = "Biometric Authentication System for FastAPI applications"
readme = "PYPI_DESCRIPTION.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Sami Melhem", email = "SaMiLMelhem23@gmail.com"}
]
keywords = ["authentication", "biometric", "fastapi", "webauthn", "fido2", "security"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Security",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Framework :: FastAPI",
]

dependencies = [
    "fastapi>=0.121.2",
    "pydantic>=2.12.4",
    "cryptography>=46.0.3",
    "python-multipart>=0.0.20",
    "httpx>=0.28.1",
    "email-validator>=2.3.0",
    "python-dotenv>=1.2.1",
    "qrcode>=8.2",
    "pillow>=12.0.0",
    "supabase>=2.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "twine>=6.0.0",
    "uvicorn[standard]>=0.24.0",
]
supabase = [
    "supabase>=2.24.0",
]

[project.scripts]
bmauth-dev-tunnel = "bmauth.dev_tunnel:main"

[project.urls]
Homepage = "https://github.com/samimelhem/bmauth"
Documentation = "https://github.com/samimelhem/bmauth#readme"
Repository = "https://github.com/samimelhem/bmauth"
"Bug Tracker" = "https://github.com/samimelhem/bmauth/issues"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true